:truck: chore(api) move event index to `src`
