+++ api/src/certification/session-management/application/jobs/certification-rescoring-by-script-job-controller.js
@@ -1,5 +1,5 @@
-import { eventDispatcher } from '../../../../../lib/domain/events/index.js';
 import { JobController } from '../../../../shared/application/jobs/job-controller.js';
+import { eventDispatcher } from '../../../../shared/domain/events/index.js';
 import CertificationRescoredByScript from '../../domain/events/CertificationRescoredByScript.js';
 import { CertificationRescoringByScriptJob } from '../../domain/models/CertificationRescoringByScriptJob.js';
