+++ api/tests/certification/session-management/unit/application/finalize-controller_test.js
@@ -1,7 +1,7 @@
-import * as events from '../../../../../lib/domain/events/index.js';
+import { DomainTransaction } from '../../../../../src/shared/domain/DomainTransaction.js';
 import { finalizeController } from '../../../../../src/certification/session-management/application/finalize-controller.js';
 import { usecases } from '../../../../../src/certification/session-management/domain/usecases/index.js';
-import { DomainTransaction } from '../../../../../src/shared/domain/DomainTransaction.js';
+import * as events from '../../../../../src/shared/domain/events/index.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Certification | Session Management | Unit | Application | Controller | Finalize', function () {
