+++ api/tests/tooling/events/event-dispatcher-builder.js
@@ -1,4 +1,4 @@
-import { _forTestOnly } from '../../../lib/domain/events/index.js';
+import { _forTestOnly } from '../../../src/shared/domain/events/index.js';
 import { sinon } from '../../test-helper.js';
 
 function buildEventDispatcherAndHandlersForTest() {
